Renamed script (TextTree) and browse function now allows only folder selection
- TextTree replaces arbo_widget_10
- Deleted arbo_widget_3
